add client-admin users & edit ACL

+++ resources/views/admin/users/index.blade.php
@@ -15,32 +15,34 @@
         <div class="panel-body table-responsive">
             <table class="table table-bordered table-striped {{ count($users) > 0 ? 'datatable' : '' }} dt-select">
                 <thead>
-                    <tr>
-                        <th style="text-align:center;"><input type="checkbox" id="select-all" /></th>
+                <tr>
+                    <th style="text-align:center;"><input type="checkbox" id="select-all"/></th>
 
-                        <th>@lang('global.users.fields.name')</th>
-                        <th>@lang('global.users.fields.email')</th>
-                        <th>@lang('global.users.fields.roles')</th>
-                        <th>&nbsp;</th>
+                    <th>@lang('global.users.fields.name')</th>
+                    <th>@lang('global.users.fields.email')</th>
+                    <th>@lang('global.users.fields.roles')</th>
+                    <th>&nbsp;</th>
 
-                    </tr>
+                </tr>
                 </thead>
-                
-                <tbody>
-                    @if (count($users) > 0)
-                        @foreach ($users as $user)
-                            <tr data-entry-id="{{ $user->id }}">
-                                <td></td>
 
-                                <td>{{ $user->name }}</td>
-                                <td>{{ $user->email }}</td>
-                                <td>
-                                    @foreach ($user->roles()->pluck('name') as $role)
-                                        <span class="label label-info label-many">{{ $role }}</span>
-                                    @endforeach
-                                </td>
-                                <td>
-                                    <a href="{{ route('admin.users.edit',[$user->id]) }}" class="btn btn-xs btn-info">@lang('global.app_edit')</a>
+                <tbody>
+                @if (count($users) > 0)
+                    @foreach ($users as $user)
+                        <tr data-entry-id="{{ $user->id }}">
+                            <td></td>
+
+                            <td>{{ $user->name }}</td>
+                            <td>{{ $user->email }}</td>
+                            <td>
+                                @foreach ($user->roles()->pluck('name') as $role)
+                                    <span class="label label-info label-many">{{ $role }}</span>
+                                @endforeach
+                            </td>
+                            <td>
+                                @if($user->id > 2 )
+                                    <a href="{{ route('admin.users.edit',[$user->id]) }}"
+                                       class="btn btn-xs btn-info">@lang('global.app_edit')</a>
                                     {!! Form::open(array(
                                         'style' => 'display: inline-block;',
                                         'method' => 'DELETE',
@@ -48,22 +50,22 @@
                                         'route' => ['admin.users.destroy', $user->id])) !!}
                                     {!! Form::submit(trans('global.app_delete'), array('class' => 'btn btn-xs btn-danger')) !!}
                                     {!! Form::close() !!}
-                                </td>
-
-                            </tr>
-                        @endforeach
-                    @else
-                        <tr>
-                            <td colspan="9">@lang('global.app_no_entries_in_table')</td>
+                                @endif
+                            </td>
                         </tr>
-                    @endif
+                    @endforeach
+                @else
+                    <tr>
+                        <td colspan="9">@lang('global.app_no_entries_in_table')</td>
+                    </tr>
+                @endif
                 </tbody>
             </table>
         </div>
     </div>
 @stop
 
-@section('javascript') 
+@section('javascript')
     <script>
         window.route_mass_crud_entries_destroy = '{{ route('admin.users.mass_destroy') }}';
     </script>
